Added Error Handling

+++ aggregate.js
@@ -5,21 +5,26 @@ const fs = require('fs');
 
 const app = express()
 
-var filePath = './templates.json'; 
-fs.unlinkSync(filePath);
+var filePath = './templates.json';
+if (filePath === true) {
+    fs.unlinkSync(filePath);
+} else {
+    "template.json not found"
+}
 
-var templateCompilation = {             
+
+var templateCompilation = {
     "version": "2",
     "templates": []
 }
 
 const templateSites = [
-        `https://raw.githubusercontent.com/portainer/templates/master/templates-2.0.json`,
-        `https://raw.githubusercontent.com/dnburgess/self-hosted-template/master/template.json`,
-        `https://raw.githubusercontent.com/Qballjos/portainer_templates/master/Template/template.json`,
-        `https://raw.githubusercontent.com/SelfhostedPro/selfhosted_templates/portainer-2.0/Template/template.json`,
-        `https://raw.githubusercontent.com/technorabilia/portainer-templates/main/lsio/templates/templates-2.0.json`,
-        `https://raw.githubusercontent.com/mikestraney/portainer-templates/master/templates.json`,
+    `https://raw.githubusercontent.com/portainer/templates/master/templates-2.0.json`,
+    `https://raw.githubusercontent.com/dnburgess/self-hosted-template/master/template.json`,
+    `https://raw.githubusercontent.com/Qballjos/portainer_templates/master/Template/template.json`,
+    `https://raw.githubusercontent.com/SelfhostedPro/selfhosted_templates/portainer-2.0/Template/template.json`,
+    `https://raw.githubusercontent.com/technorabilia/portainer-templates/main/lsio/templates/templates-2.0.json`,
+    `https://raw.githubusercontent.com/mikestraney/portainer-templates/master/templates.json`,
 ]
 
 app.use(express.static('public'));
@@ -28,23 +33,27 @@ app.use(express.static('public'));
 const sleep = ms => new Promise(res => setTimeout(res, ms))
 
 const compileTemplates = () => new Promise((resolve, reject) => {
-    templateCompilation = {             
-        "version": "2",
-        "templates": []
-    }
- templateSites.map(templates =>
-        request.get(templates).then(response => {
-            
-            for (let i in response.data.templates){
-                templateCompilation['templates'].push(response.data.templates[i])
-            }
+    try {
+        templateCompilation = {
+            "version": "2",
+            "templates": []
+        }
+        templateSites.map(templates =>
+            request.get(templates).then(response => {
 
-            resolve(response)
-            return templateCompilation
-        })
-    ) 
-    
-    
+                for (let i in response.data.templates) {
+                    templateCompilation['templates'].push(response.data.templates[i])
+                }
+
+                resolve(response)
+                return templateCompilation
+            })
+        )
+
+    } catch (err) {
+        reject(`Error during get request: ${err}`);
+    }
+    return
 })
 
 const printObject = async () => {
